+++ src/types/mypage.ts
@@ -1,0 +1,7 @@
+export interface MoodBoardData {
+  id: number;
+  title: string;
+  category: string;
+  lock: boolean;
+  profileNum: number;
+}
